requirements.txt
python-dotenv==1.0.0
transformers==4.33.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
requests==2.31.0